# docs/api.yml
openapi: 3.0.3
info:
  title: 花道展写真共有用アルバム
  description: 花道展の作品や花展の情報を管理するためのAPIです。
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
    description: ローカル環境
paths:
  /exhibitions:
    get:
      tags:
        - exhibition
      summary: 花展の一覧
      description: 花展の一覧を取得する。花展は開催日の降順。
      operationId: listExhibitions
      responses:
        '200':
          $ref: '#/components/responses/ExhibitionsResponse'
        default:
          $ref: '#/components/responses/UnexpectedError'
    post:
      tags:
        - exhibition
      summary: 花展の登録
      description: 花展を登録する
      operationId: createExhibition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExhibitionPayload'
      responses:
        '201':
          $ref: '#/components/responses/CreatedExhibitionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /exhibitions/{exhibitionId}:
    get:
      tags:
        - exhibition
      summary: 花展の取得
      description: 花展の情報を取得する
      operationId: getExhibition
      parameters:
        - $ref: '#/components/parameters/ExhibitionIdParam'
      responses:
        '200':
          $ref: '#/components/responses/ExhibitionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags:
        - exhibition
      summary: 花展の更新
      description: 花展の情報を更新する
      operationId: updateExhibition
      parameters:
        - $ref: '#/components/parameters/ExhibitionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExhibitionPayload'
      responses:
        '204':
          description: 花展が更新された
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
        - exhibition
      summary: 花展の削除
      description: 花展を削除する
      operationId: deleteExhibition
      parameters:
        - $ref: '#/components/parameters/ExhibitionIdParam'
      responses:
        '204':
          description: 花展が削除された
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /exhibitions/{exhibitionId}/works:
    get:
      tags:
        - exhibition
      summary: 花展の作品の一覧
      description: 花展ごとの作品の一覧を登録の作品の登録日の昇順で取得する
      operationId: listExhibitionWorks
      parameters:
        - $ref: '#/components/parameters/ExhibitionIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error
  /exhibitions/{exhibitionId}/works/{workId}:
    get:
      tags:
        - exhibition
      summary: 花展の作品の取得
      description: 花展ごとの作品の情報を取得する
      operationId: getExhibitionWork
      parameters:
        - $ref: '#/components/parameters/ExhibitionIdParam'
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error
...
  /arrangers/{arrangerId}:
    get:
      tags:
        - arranger
      summary: 作者の情報の取得
      description: 作者の情報を取得する
      operationId: getArranger
      parameters:
        - $ref: '#/components/parameters/ArrangerIdParam'
      responses:
        '200':
          $ref: '#/components/responses/ArrangerResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error
    put:
      tags:
        - arranger
      summary: 作者の更新
      description: 作者の情報を更新する
      operationId: updateArranger
      parameters:
        - $ref: '#/components/parameters/ArrangerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArrangerPayload'
      responses:
        '204':
          description: 作者が更新された
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
        - arranger
      summary: 作者の削除
      description: 作者を削除する
      operationId: deleteArranger
      parameters:
        - $ref: '#/components/parameters/ArrangerIdParam'
      responses:
        '204':
          description: 作者が削除された
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /arrangers/{arrangerId}/works:
    get:
      tags:
        - arranger
      summary: 作者の作品の一覧
      description: 作者ごとの作品の一覧を作品の登録日の昇順で取得する
      operationId: listArrangerWorks
      parameters:
        - $ref: '#/components/parameters/ArrangerIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /arrangers/{arrangerId}/works/{workId}:
    get:
      tags:
        - arranger
      summary: 作者の作品の取得
      description: 作者ごとの作品の情報を取得する
      operationId: getArrangerWork
      parameters:
        - $ref: '#/components/parameters/ArrangerIdParam'
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /materials/{materialId}:
    get:
      tags:
        - material
      summary: 材料の情報の取得
      description: 材料の情報を取得する
      operationId: getMaterial
      parameters:
        - $ref: '#/components/parameters/MaterialIdParam'
      responses:
        '200':
          $ref: '#/components/responses/MaterialResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags:
        - material
      summary: 材料の更新
      description: 材料の情報を更新する
      operationId: updateMaterial
      parameters:
        - $ref: '#/components/parameters/MaterialIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaterialPayload'
      responses:
        '204':
          description: 材料が更新された
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
        - material
      summary: 材料の削除
      description: 材料を削除する
      operationId: deleteMaterial
      parameters:
        - $ref: '#/components/parameters/MaterialIdParam'
      responses:
        '204':
          description: 材料が削除された
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /materials/{materialId}/works:
    get:
      tags:
        - material
      summary: 材料の作品の一覧
      description: 材料ごとの作品の一覧を作品の登録日の昇順で取得する
      operationId: listMaterialWorks
      parameters:
        - $ref: '#/components/parameters/MaterialIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /materials/{materialId}/works/{workId}:
    get:
      tags:
        - material
      summary: 材料の作品の取得
      description: 材料ごとの作品の情報を取得する
      operationId: getMaterialWork
      parameters:
        - $ref: '#/components/parameters/MaterialIdParam'
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /categories/{categoryId}:
    get:
      tags:
        - category
      summary: カテゴリの情報の取得
      description: カテゴリの情報を取得する
      operationId: getCategory
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          $ref: '#/components/responses/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags:
        - category
      summary: カテゴリの更新
      description: カテゴリの情報を更新する
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryPayload'
      responses:
        '204':
          description: カテゴリが更新された
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
        - category
      summary: カテゴリの削除
      description: カテゴリを削除する
      operationId: deleteCategory
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '204':
          description: カテゴリが削除された
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /categories/{categoryId}/works:
    get:
      tags:
        - category
      summary: カテゴリの作品の一覧
      description: カテゴリごとの作品の一覧を作品の登録日の昇順で取得する
      operationId: listCategoryWorks
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /categories/{categoryId}/works/{workId}:
    get:
      tags:
        - category
      summary: カテゴリの作品の取得
      description: カテゴリごとの作品の情報を取得する
      operationId: getCategoryWork
      parameters:
        - $ref: '#/components/parameters/CategoryIdParam'
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /seasons/{seasonId}:
    get:
      tags:
        - season
      summary: 季節の情報の取得
      description: 季節の情報を取得する
      operationId: getSeason
      parameters:
        - $ref: '#/components/parameters/SeasonIdParam'
      responses:
        '200':
          $ref: '#/components/responses/SeasonResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    put:
      tags:
        - season
      summary: 季節の更新
      description: 季節の情報を更新する
      operationId: updateSeason
      parameters:
        - $ref: '#/components/parameters/SeasonIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Season'
      responses:
        '204':
          description: 季節が更新された
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      tags:
        - season
      summary: 季節の削除
      description: 季節を削除する
      operationId: deleteSeason
      parameters:
        - $ref: '#/components/parameters/SeasonIdParam'
      responses:
        '204':
          description: 季節が削除された
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /seasons/{seasonId}/works:
    get:
      tags:
        - season
      summary: 季節の作品の一覧
      description: 季節ごとの作品の一覧を作品の登録日の昇順で取得する
      operationId: listSeasonWorks
      parameters:
        - $ref: '#/components/parameters/SeasonIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /seasons/{seasonId}/works/{workId}:
    get:
      tags:
        - season
      summary: 季節の作品の取得
      description: 季節ごとの作品の情報を取得する
      operationId: getSeasonWork
      parameters:
        - $ref: '#/components/parameters/SeasonIdParam'
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          $ref: '#/components/responses/WorkListItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  schemas:
...
    CreateExhibitionPayload:
      type: object
      properties:
        name:
          type: string
          description: 花展の名前
        started_date:
          type: string
          format: date
          description: 花展の開始日
        ended_date:
          type: string
          format: date
          description: 花展の終了日
      required:
        - name
        - started_date
        - ended_date
      example:
        name: "春の花展"
        started_date: "2023-03-01"
        ended_date: "2023-03-31"
    UpdateExhibitionPayload:
      type: object
      properties:
        name:
          type: string
          description: 花展の名前
        started_date:
          type: string
          format: date
          description: 花展の開始日
        ended_date:
          type: string
          format: date
          description: 花展の終了日
      required:
        - name
        - started_date
        - ended_date
      example:
        name: "春の花展"
        started_date: "2023-03-01"
        ended_date: "2023-03-31"
...
    UpdateArrangerPayload:
      type: object
      properties:
        name:
          type: string
          description: 作者の名前
      required:
        - name
      example:
        name: "山田太郎"
...
  parameters:
    ExhibitionIdParam:
      name: exhibitionId
      in: path
      required: true
      description: 花展のID
      schema:
        type: integer
    WorkIdParam:
      name: workId
      in: path
      required: true
      description: 作品のID
      schema:
        type: integer
    ArrangerIdParam:
      name: arrangerId
      in: path
      required: true
      description: 作者のID
      schema:
        type: integer
...
  responses:
    UnexpectedError:
      description: An unexpected error occurred
    NotFound:
      description: データが見つからない
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "データが見つかりません"
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "入力内容が正しくありません"
    ExhibitionResponse:
      description: 単一の花展の情報
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Exhibition'
    ExhibitionsResponse:
      description: 花展の一覧
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Exhibition'
            example:
              - id: 1
                name: "春の花展"
                started_date: "2023-03-01"
                ended_date: "2023-03-31"
              - id: 2
                name: "夏の花展"
                started_date: "2023-06-01"
                ended_date: "2023-06-30"
              - id: 3
                name: "秋の花展"
                started_date: "2023-09-01"
                ended_date: "2023-09-30"
              - id: 4
                name: "冬の花展"
                started_date: "2023-12-01"
                ended_date: "2023-12-31"
...
    WorkListItemResponse:
      description: 花展の作品一覧の中の単一の作品
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkListItem'
...
    ArrangerResponse:
      description: 単一の作者
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Arranger'
...
    CreatedExhibitionResponse:
      description: 花展の作成の成功
      headers:
        Location:
          description: 作成された花展のパス
          schema:
            type: string
            example: "/exhibitions/1"
...